+++ src/models/Settings.php
@@ -2,79 +2,91 @@
 
 namespace codemonauts\instagramfeed\models;
 
-use codemonauts\instagramfeed\InstagramFeed;
 use Craft;
 use craft\base\Model;
+use craft\helpers\App;
 
 class Settings extends Model
 {
     /**
      * @var string The Instagram account to get the feed from
      */
-    public $instagramUser = '';
+    public string $instagramUser = '';
 
     /**
      * @var boolean Use codemonauts proxy to get the Instagram page
      */
-    public $useProxy = false;
+    public bool $useProxy = false;
 
     /**
      * @var string The proxy key to use for authentication
      */
-    public $proxyKey = '';
+    public string $proxyKey = '';
 
     /**
      * @var int Timeout in seconds waiting for the Instagram page to load
      */
-    public $timeout = 10;
-
-    /**
-     * @deprecated Guzzle will always be used.
-     * @var boolean Use Guzzle instead of php's file stream to fetch the Instagram page
-     */
-    public $useGuzzle = false;
+    public int $timeout = 10;
 
     /**
      * @var boolean Dump Instagram response in file for debugging purpose
      */
-    public $dump = false;
-
-    /**
-     * @var string User Agent to use when fetching the Instagram page
-     */
-    public $userAgent = 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_14_6) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/77.0.3865.75 Safari/537.36';
+    public bool $dump = false;
 
     /**
      * @var boolean Use volumes to store Instagram images
      */
-    public $useVolume = false;
+    public bool $useVolume = false;
 
     /**
      * @var string The volume handle to use for storing Instagram images
      */
-    public $volume = '';
+    public string $volume = '';
 
     /**
      * @var string Subpath to use on volumes to store Instagram images
      */
-    public $subpath = '';
+    public string $subpath = '';
 
     /**
      * @inheritdoc
      */
-    public function rules()
+    public function rules(): array
     {
         return [
-            [['instagramUser', 'timeout', 'userAgent'], 'required'],
+            [['instagramUser', 'timeout'], 'required'],
+            ['instagramUser', 'validateInstagramUser'],
             ['timeout', 'double', 'min' => 1],
-            [['useProxy', 'useVolume'], 'boolean'],
-            ['dump', 'boolean'],
-            [['userAgent', 'proxyKey', 'volume', 'subpath'], 'string'],
+            [['useProxy', 'useVolume', 'dump'], 'boolean'],
+            [['proxyKey', 'volume', 'subpath'], 'string'],
         ];
     }
 
-    public function getAccount()
+    /**
+     * Validator for Instagram account names
+     *
+     * @param string $attribute
+     */
+    public function validateInstagramUser(string $attribute): void
     {
-        return Craft::parseEnv(trim(InstagramFeed::getInstance()->getSettings()->instagramUser));
+        $value = $this->$attribute;
+
+        if (str_starts_with($value, '#') || str_starts_with($value, '$')) {
+            return;
+        }
+
+        if (!preg_match('/^[a-zA-Z0-9._]{1,30}$/', $value)) {
+            $this->addError($attribute, Craft::t('instagramfeed', 'Not a valid Instagram account name.'));
+        }
+    }
+
+    /**
+     * Returns the parsed account.
+     *
+     * @return string
+     */
+    public function getAccount(): string
+    {
+        return App::parseEnv(trim($this->instagramUser));
     }
 }