+++ src/translations/de/instagramfeed.php
@@ -1,8 +1,12 @@
 <?php
 
 return [
-	// Settings
-	'Username' => 'Benutzername',
-	'Please, enter your Instagram username.' => 'Bitte geben Sie den Instagram Benutzername ein.',
-	'Make sure the entered Instagram user profile is not private.' => 'Stellen Sie sicher, dass der angegebene Benutzer auch ein öffentliches Profil besitzt.',
+    // Settings
+    'Username' => 'Benutzername',
+    'Please, enter your Instagram username.' => 'Bitte geben Sie den Instagram Benutzername ein.',
+    'Make sure the entered Instagram user profile is not private.' => 'Stellen Sie sicher, dass der angegebene Benutzer auch ein öffentliches Profil besitzt.',
+    'Not a valid Instagram account name.' => 'Kein gültiger Instagram Account Name.',
+    'This is being overridden by the config file.' => 'Dies wird von der Konfigurationsdatei überschrieben.',
+    'Proxy usage is enabled but the proxy key is empty, so the proxy will not be used.' => 'Proxy ist aktiviert, aber der Access Key fehlt. Daher wird er nicht verwendet.',
+    'Proxy will be used.' => 'Proxy wird verwendet.',
 ];